Use $(PKG_CONFIG) instead of pkg-config
On cross-compilation the pkg-config binary is usually substituted by a
wrapper (e.g. i686-w64-mingw32-pkg-config) that provides the proper
environment to the original pkg-config tool. Ensure to use this wrapper
when required: it is set in the PKG_CONFIG variable by configure.

On straight build PKG_CONFIG is set to pkg-config, so backward
compatibility is retained.